requirements.txt
mypy
marshmallow
marshmallow-dataclass
pydantic
pytypes
typical